requirements/infer.txt
grpcio==1.26.0
grpcio-tools==1.26.0
mtcnn==0.1.0
Pillow==7.0.0
pytz==2019.3